fixed several warnings, got Header proto sending from android to the server

+++ AndroidClientProject/AndroidClient/src/main/java/com/example/myfirstapp/MusicListActivity.java
@@ -21,7 +21,7 @@ public class MusicListActivity extends BaseActivity {
                     .build();
 
             byte[] data = header.toByteArray();
-            short length = (short) data.length;
+            int length = data.length;
 
             mBoundService.sendPrefixLength(length);
             mBoundService.sendMessage(data);
